backlight: lm3697: Add lm3697 backlight driver
This driver is a general version for LM3697 backlgiht + flash driver chip
of TI.

LM3697:
The LM3697 is a high-efficiency three-string power source for backlight
or keypad LEDs in smart-phone handsets. The high-voltage inductive boost
converter provides the power for three-series LED strings for display
backlight and keypad functions (HVLED1, HVLED2, and HVLED3).